finance-prediction-system/backend/data_processor/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
psycopg2-binary==2.9.5
redis==4.5.1
pandas==1.5.3
numpy==1.24.3
python-dotenv==1.0.0
sqlalchemy==2.0.9
